+++ HISTORY.rst
@@ -6,6 +6,7 @@ History
 v0.16.0a1 (2010-03-06)
 ......................
 * add ``job_try`` argument to ``enqueue_job``, #113
+* adding ``--watch`` mode to the worker (requires ``watchgod``), #114
 
 v0.16.0a1 (2010-03-05)
 ......................